feat: Added re-order and move attributes [PT-186024265] [PT-186024290]
This adds a draggable table context with a draggable header element that allows users to re-order columns in the collection and move elements to new collections.

This also has some existing missing key cleanup work in it.

NOTE: This does not use DnDKit as that does not work with tables.